fix: buildできるように修正
